add readme to the scripty

+++ scripts/build_comfyui_desktop.cmd
@@ -15,20 +15,29 @@ set TORCHAUDIO_VERSION=2.7
 set TORCHVISION_VERSION=0.22
 set CUDA_PIP_INDEX=cu128
 
+REM path to node version manager. Change it if you installed it somewhere else.
+set NVM_HOME=%LocalAppData%\nvm
+
 set PYTHON_VERSION_STR=%PYTHON_VERSION:.=%
 
 set WORK_DIR=%cd%
 
-REM 1. Install Python 3.12 silently with winget
-echo Installing Python %PYTHON_VERSION%...
-winget install -e --id Python.Python.%PYTHON_VERSION% --accept-source-agreements --accept-package-agreements -h
-if %errorlevel% neq 0 (
-    echo Failed to install Python %PYTHON_VERSION%
-    exit /b 1
+REM Assume Python 3.12 installs here. This is the default location for winget installations. 
+REM Adjust if your installation path is different.
+set PYTHON_EXE="%LocalAppData%\Programs\Python\Python%PYTHON_VERSION_STR%\python.exe"
+
+REM 1. Install Python 3.12 silently with winget, skip if PYTHON_EXE already exists
+if exist %PYTHON_EXE% (
+    echo Python %PYTHON_VERSION% is already installed. Skip downloading..
+) else (
+    echo Installing Python %PYTHON_VERSION%...
+    winget install -e --id Python.Python.%PYTHON_VERSION% --accept-source-agreements --accept-package-agreements -h
+    if %errorlevel% neq 0 (
+        echo Failed to install Python %PYTHON_VERSION%
+        exit /b 1
+    )
 )
 
-REM Assume Python 3.12 installs here:
-set PYTHON_EXE="%LocalAppData%\Programs\Python\Python%PYTHON_VERSION_STR%\python.exe"
 
 REM 2. Install uv package
 echo Installing uv package...
@@ -45,24 +54,22 @@ if %errorlevel% neq 0 (
 
 REM 4. Install Node.js 20 via NVM
 echo Installing Node.js %NODE_VERSION% with NVM...
-set NVM_HOME=%LocalAppData%\nvm
 cd %NVM_HOME%
 nvm install %NODE_VERSION%
 nvm use %NODE_VERSION%
 
-REM 5. Install Yarn using npm
+REM 5. Clone ComfyUI desktop repo
+echo Cloning ComfyUI Desktop...
+cd %WORK_DIR%
+git clone https://github.com/kungtalon/desktop.git
+cd desktop
+
+REM 6. Install Yarn using npm
 REM Note: this step needs admin permission
 echo Installing yarn...
 npm install -g yarn
 corepack enable
 yarn use %YARN_VERSION%
-
-REM 6. Clone ComfyUI desktop repo
-echo Cloning ComfyUI Desktop...
-cd %WORK_DIR%
-git clone https://github.com/Comfy-Org/desktop.git
-cd desktop
-git clone https://github.com/nunchaku-tech/ComfyUI-nunchaku.git assets/ComfyUI/custom_nodes/ComfyUI-nunchaku
 
 REM 7. Install node modules and rebuild electron
 echo Rebuilding native modules...
@@ -72,6 +79,8 @@ yarn make:assets
 
 REM 8. Overwrite override.txt with torch 2.7 + custom nunchaku wheel
 echo Writing override.txt...
+
+git clone https://github.com/nunchaku-tech/ComfyUI-nunchaku.git assets/ComfyUI/custom_nodes/ComfyUI-nunchaku
 
 set NUNCHAKU_URL=https://github.com/nunchaku-tech/nunchaku/releases/download/v%NUNCHAKU_VERSION%/nunchaku-%NUNCHAKU_VERSION%+torch%TORCH_VERSION%-cp%PYTHON_VERSION_STR%-cp%PYTHON_VERSION_STR%-win_amd64.whl
